requirements.txt
asgiref==3.5.2
Django==4.0.4
djangorestframework==3.13.1
mysql==0.0.3
mysqlclient==2.1.0
pytz==2022.1
Serializer==0.2.1
simplejson==3.17.6
sqlparse==0.4.2
tzdata==2022.1